inverted_index_ffi:: split inverted_index.rs into submodules
