+++ src/redisearch_rs/c_entrypoint/iterators_ffi/src/inverted_index/wildcard.rs
@@ -1,0 +1,177 @@
+/*
+ * Copyright (c) 2006-Present, Redis Ltd.
+ * All rights reserved.
+ *
+ * Licensed under your choice of the Redis Source Available License 2.0
+ * (RSALv2); or (b) the Server Side Public License v1 (SSPLv1); or (c) the
+ * GNU Affero General Public License v3 (AGPLv3).
+*/
+
+use std::{fmt::Debug, ptr::NonNull};
+
+use inverted_index::{
+    IndexReader, RSIndexResult, doc_ids_only::DocIdsOnly, raw_doc_ids_only::RawDocIdsOnly, t_docId,
+};
+use rqe_iterators::inverted_index::Wildcard;
+use rqe_iterators_interop::RQEIteratorWrapper;
+
+/// Wrapper around different II wildcard iterator encoding types to avoid generics in FFI code.
+///
+/// Handles both the standard variable-length encoding ([`DocIdsOnly`]) and the
+/// fixed 4-byte raw encoding ([`RawDocIdsOnly`]).
+pub(super) enum WildcardIterator<'index> {
+    Encoded(Wildcard<'index, DocIdsOnly>),
+    Raw(Wildcard<'index, RawDocIdsOnly>),
+}
+
+impl Debug for WildcardIterator<'_> {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        let variant = match self {
+            WildcardIterator::Encoded(_) => "Encoded",
+            WildcardIterator::Raw(_) => "Raw",
+        };
+        write!(f, "WildcardIterator({variant})")
+    }
+}
+
+impl WildcardIterator<'_> {
+    /// Get the flags from the underlying reader.
+    pub(super) fn flags(&self) -> ffi::IndexFlags {
+        match self {
+            WildcardIterator::Encoded(w) => w.reader().flags(),
+            WildcardIterator::Raw(w) => w.reader().flags(),
+        }
+    }
+}
+
+impl<'index> rqe_iterators::RQEIterator<'index> for WildcardIterator<'index> {
+    #[inline(always)]
+    fn current(&mut self) -> Option<&mut RSIndexResult<'index>> {
+        match self {
+            WildcardIterator::Encoded(w) => w.current(),
+            WildcardIterator::Raw(w) => w.current(),
+        }
+    }
+
+    #[inline(always)]
+    fn read(
+        &mut self,
+    ) -> Result<Option<&mut RSIndexResult<'index>>, rqe_iterators::RQEIteratorError> {
+        match self {
+            WildcardIterator::Encoded(w) => w.read(),
+            WildcardIterator::Raw(w) => w.read(),
+        }
+    }
+
+    #[inline(always)]
+    fn skip_to(
+        &mut self,
+        doc_id: t_docId,
+    ) -> Result<Option<rqe_iterators::SkipToOutcome<'_, 'index>>, rqe_iterators::RQEIteratorError>
+    {
+        match self {
+            WildcardIterator::Encoded(w) => w.skip_to(doc_id),
+            WildcardIterator::Raw(w) => w.skip_to(doc_id),
+        }
+    }
+
+    #[inline(always)]
+    fn rewind(&mut self) {
+        match self {
+            WildcardIterator::Encoded(w) => w.rewind(),
+            WildcardIterator::Raw(w) => w.rewind(),
+        }
+    }
+
+    #[inline(always)]
+    fn num_estimated(&self) -> usize {
+        match self {
+            WildcardIterator::Encoded(w) => w.num_estimated(),
+            WildcardIterator::Raw(w) => w.num_estimated(),
+        }
+    }
+
+    #[inline(always)]
+    fn last_doc_id(&self) -> t_docId {
+        match self {
+            WildcardIterator::Encoded(w) => w.last_doc_id(),
+            WildcardIterator::Raw(w) => w.last_doc_id(),
+        }
+    }
+
+    #[inline(always)]
+    fn at_eof(&self) -> bool {
+        match self {
+            WildcardIterator::Encoded(w) => w.at_eof(),
+            WildcardIterator::Raw(w) => w.at_eof(),
+        }
+    }
+
+    #[inline(always)]
+    fn revalidate(
+        &mut self,
+    ) -> Result<rqe_iterators::RQEValidateStatus<'_, 'index>, rqe_iterators::RQEIteratorError> {
+        match self {
+            WildcardIterator::Encoded(w) => w.revalidate(),
+            WildcardIterator::Raw(w) => w.revalidate(),
+        }
+    }
+}
+
+/// Creates a new wildcard inverted index iterator for querying all existing documents.
+///
+/// # Parameters
+///
+/// * `idx` - Pointer to the existingDocs inverted index (DocIdsOnly or RawDocIdsOnly encoded).
+/// * `sctx` - Pointer to the Redis search context.
+/// * `weight` - Weight to apply to all results.
+///
+/// # Returns
+///
+/// A pointer to a `QueryIterator` that can be used from C code.
+///
+/// # Safety
+///
+/// The following invariants must be upheld when calling this function:
+///
+/// 1. `idx` must be a valid pointer to an `InvertedIndex` and cannot be NULL.
+/// 2. `idx` must remain valid between `revalidate()` calls, since the revalidation
+///    mechanism detects when the index has been replaced via `spec.existingDocs` pointer
+///    comparison.
+/// 3. `sctx` must be a valid pointer to a `RedisSearchCtx` and cannot be NULL.
+/// 4. `sctx` and `sctx.spec` must remain valid for the lifetime of the returned iterator.
+#[unsafe(no_mangle)]
+pub unsafe extern "C" fn NewInvIndIterator_WildcardQuery(
+    idx: *const ffi::InvertedIndex,
+    sctx: *const ffi::RedisSearchCtx,
+    weight: f64,
+) -> *mut ffi::QueryIterator {
+    debug_assert!(!idx.is_null(), "idx must not be null");
+
+    // Cast to the FFI wrapper enum which handles type dispatch
+    let idx_ffi: *const inverted_index_ffi::InvertedIndex = idx.cast();
+    // SAFETY: 1. guarantees idx is valid and non-null
+    let ii_ref = unsafe { &*idx_ffi };
+
+    debug_assert!(!sctx.is_null(), "sctx must not be null");
+    // SAFETY: 3. guarantees sctx is valid and non-null
+    let sctx = unsafe { NonNull::new_unchecked(sctx as *mut _) };
+
+    // Create the appropriate wildcard iterator variant based on the encoding type
+    let iterator = match ii_ref {
+        inverted_index_ffi::InvertedIndex::DocIdsOnly(ii) => {
+            // SAFETY: 3. and 4. guarantee `sctx` and `sctx.spec` validity for the iterator's lifetime.
+            WildcardIterator::Encoded(unsafe { Wildcard::new(ii.reader(), sctx, weight) })
+        }
+        inverted_index_ffi::InvertedIndex::RawDocIdsOnly(ii) => {
+            // SAFETY: 3. and 4. guarantee `sctx` and `sctx.spec` validity for the iterator's lifetime.
+            WildcardIterator::Raw(unsafe { Wildcard::new(ii.reader(), sctx, weight) })
+        }
+        _ => panic!(
+            "Wildcard iterator requires a DocIdsOnly or RawDocIdsOnly inverted index, got: {:?}",
+            std::mem::discriminant(ii_ref)
+        ),
+    };
+
+    RQEIteratorWrapper::boxed_new(ffi::IteratorType_INV_IDX_WILDCARD_ITERATOR, iterator)
+}
